refactor: instructor class URL 수정
