+++ just-swim-app/src/app/(afterLogin)/instructor/class/_components/listContent/listContent.module.scss
