+++ just-swim-app/src/app/(afterLogin)/instructor/class/_components/search/search.module.scss
