+++ just-swim-app/src/app/(afterLogin)/instructor/class/customerDetail/_components/customerClass/page.tsx
