+++ just-swim-app/src/app/(afterLogin)/instructor/class/customerDetail/_components/customerInfo/customerInfo.module.scss
