+++ just-swim-app/src/app/(afterLogin)/instructor/class/customerDetail/page.tsx
