+++ just-swim-app/src/app/(afterLogin)/instructor/class/detail/[id]/classDetail.module.scss
