+++ just-swim-app/src/app/(afterLogin)/instructor/class/detail/[id]/members/page.tsx
