+++ just-swim-app/src/app/(afterLogin)/instructor/class/detail/[id]/page.tsx
