+++ just-swim-app/src/app/(afterLogin)/instructor/class/edit/[id]/classInfoEdit.module.scss
