+++ just-swim-app/src/app/(afterLogin)/instructor/class/edit/[id]/page.tsx
