+++ just-swim-app/src/app/(afterLogin)/instructor/class/edit/[id]/search/locationSearch.module.scss
