--- just-swim-app/src/app/(afterLogin)/instructor/classList/classListTabs/_components/tabContent/tabContent.scss
@@ -1,60 +1,0 @@
-@import '@/globals.scss';
-
-.title {
-  @include Title3-M;
-  margin: 22px 16px 12px 16px;
-}
-
-.tabList {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  justify-items: center;
-  margin: 16px;
-}
-
-.tabContent {
-  width: 158px;
-  height: fit-content;
-  background-color: var(--gray-50);
-  border-radius: 13px;
-  padding: 12px;
-  margin-bottom: 12px;
-}
-
-.bar {
-  width: 100vw;
-  height: 8px;
-  background-color: var(--gray-50);
-}
-
-.icon {
-  margin-right: 6px;
-}
-
-.name {
-  @include Title3-SB;
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 4px;
-}
-
-.info {
-  @include Body2-R;
-  margin: 18px 0;
-}
-
-.profileBox {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  .photoList {
-    img:nth-child(n + 2) {
-      margin-left: -8px;
-    }
-
-    p {
-      @include Title5-SB;
-    }
-  }
-}
